+++ packages/testpkg2/CHANGELOG.md
@@ -1,5 +1,25 @@
 # @prosopo/testpkg2
 
+## 1.0.12
+### Patch Changes
+
+- 6fe8570: Release 3.2.0
+- Updated dependencies [6fe8570]
+  - @prosopo/config@3.1.8
+
+## 1.0.11
+### Patch Changes
+
+- f304be9: Release 3.1.13
+- Updated dependencies [f304be9]
+  - @prosopo/config@3.1.7
+
+## 1.0.10
+### Patch Changes
+
+- Updated dependencies [9eed772]
+  - @prosopo/config@3.1.6
+
 ## 1.0.9
 ### Patch Changes